app/requirements.txt
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
email-validator==2.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
cryptography==42.0.5
PyMySQL==1.1.0
Werkzeug==3.0.1
alembic==1.13.1
itsdangerous==2.1.2
Jinja2==3.1.3
SQLAlchemy==2.0.28
gunicorn==21.2.0
pytest==8.0.2
black==24.2.0
flake8==7.0.0

